requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
azure-storage-queue
pytest
coverage